examples/django/requirements.txt
django==2.1.5
djangorestframework==3.9.1
django-dotenv==1.4.2
https://github.com/apixu/apixu-python/archive/vX.X.X.zip
